requirements.txt
amqp==5.3.1
boto3==1.42.88
botocore==1.42.88
jmespath==1.1.0
s3transfer==0.16.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.4
django-cors-headers==4.7.0
django-admin-extra-buttons==2.1.2
django-guardian==3.3.1
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
stripe==15.0.1
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
